smdk4412-common: enable boeffla_sound and set the speaker_volume to max

+++ configs/restart_rild.sh
@@ -1,8 +1,8 @@
 #!/system/bin/sh
 
-#set -x
-
-reset_threshold=5
+simState=$(getprop gsm.sim.state)
+bootCompleted=$(getprop sys.boot_completed)
+setprop gsm.resetreason $simState
 
 do_reset_radio() {
   resetCount=$(getprop gsm.resetcount)
@@ -10,40 +10,29 @@ do_reset_radio() {
     resetCount=0
   fi
 
-  # Allow up to one radio reset in airplane mode
-  airplane_mode=$(settings get global airplane_mode_on)
-  if [ "$airplane_mode" == "1" ] && [ $resetCount -ge 1 ] ; then
-    return;
+  if [ "$bootCompleted" == 1 ] ; then
+    if [ "$simState" == "UNKNOWN" ] ; then
+      setprop gsm.sim.state "RESET RILD"
+      PID=$(ps -A | grep "rild" | awk '{ print $2 }')
+      kill $PID
+      setprop gsm.resetcount $(( $resetCount + 1 ))
+    fi
   fi
-
-  # No radio after several resets?
-  # Probably something goes wrong, we have nothing to do.
-  if [ $resetCount -ge $reset_threshold ] ; then
-    return
+  sleep 10
+  simState=$(getprop gsm.sim.state)
+  if [ "$simState" != "LOADED" ] ; then
+    if [ "$simState" != "READY" ] ; then
+      setprop gsm.sim.state "RESET_QMUXD"
+      PIDQ=$(ps -A | grep "qmuxd" | awk '{ print $2 }')
+      kill $PIDQ
+      setprop gsm.resetcount $(( $resetCount + 1 ))
+    fi
   fi
-
-  #status=$(getprop gsm.status)
-
-  subId=$(getprop gsm.subid)
-  simState=$(getprop gsm.sim_state)
-
   sleep 30
-  simState=$(getprop gsm.sim_state)
-  if [ "$simState" != "LOADED" ] ; then
-    PID=$(ps -A | grep "rild" | awk '{ print $2 }')
-    kill $PID
-    sleep 5
-    setprop gsm.resetcount $(( $resetCount + 1 ))
+  simState=$(getprop gsm.sim.state)
+  if [ "$simState" == "RESET_QMUXD" ] ; then
+    setprop gsm.sim.state "UNKNOWN"
   fi
-  sleep 5
-  simState=$(getprop gsm.sim_state)
-  if [ "$simState" != "LOADED" ] ; then
-    PIDQ=$(ps -A | grep "qmuxd" | awk '{ print $2 }')
-    kill $PIDQ
-    sleep 5
-  fi
-
-  setprop gsm.radioreset false
 
 }
